requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
numpy==1.22.3
Pillow==9.1.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing_extensions==4.1.1
urllib3==1.26.9
